+++ lib/src/app/views/evenements/evenement_create_view.dart
@@ -7,13 +7,12 @@ import 'package:jdr_maker/src/app/tools/firebase_android_tool.dart';
 import 'package:jdr_maker/src/app/tools/firebase_desktop_tool.dart';
 import 'package:jdr_maker/src/app/tools/get_random_string.dart';
 import 'package:jdr_maker/src/app/widgets/boutons/bouton.dart';
-import 'package:jdr_maker/src/app/widgets/boutons/form_bouton.dart';
+import 'package:jdr_maker/src/app/widgets/boutons/icone_bouton.dart';
 import 'package:jdr_maker/src/app/widgets/champs/champ_saisie.dart';
 import 'package:jdr_maker/src/app/widgets/chargement.dart';
 import 'package:jdr_maker/src/app/widgets/entete_application.dart';
 import 'package:jdr_maker/src/app/widgets/interfaces/app_interface/app_interface.dart';
 import 'package:jdr_maker/src/domain/data/couleurs.dart';
-import 'package:jdr_maker/src/domain/enums/form_bouton_type.dart';
 import 'package:jdr_maker/src/domain/models/evenement_model.dart';
 import 'package:provider/provider.dart';
 
@@ -150,10 +149,12 @@ class _EvenementCreateViewState extends State<EvenementCreateView> {
                       Column(children: listeObjets()),
                     ],
                   ),
-                  FormBouton(
-                    boutonType: FormBoutonType.valider,
-                    alignement: Alignment.bottomRight,
-                    action: creer,
+                  Align(
+                    alignment: Alignment.bottomRight,
+                    child: BoutonIcone(
+                      icone: Icons.done_rounded,
+                      action: creer,
+                    ),
                   ),
                 ],
               ),