+++ lib/src/app/views/evenements/evenement_edit_view.dart
@@ -3,13 +3,12 @@ import 'package:jdr_maker/src/app/controllers/evenement_controller.dart';
 import 'package:jdr_maker/src/app/controllers/navigation_controller.dart';
 import 'package:jdr_maker/src/app/controllers/projet_controller.dart';
 import 'package:jdr_maker/src/app/tools/firebase_global_tool.dart';
-import 'package:jdr_maker/src/app/widgets/boutons/form_bouton.dart';
+import 'package:jdr_maker/src/app/widgets/boutons/icone_bouton.dart';
 import 'package:jdr_maker/src/app/widgets/champs/champ_saisie.dart';
 import 'package:jdr_maker/src/app/widgets/chargement.dart';
 import 'package:jdr_maker/src/app/widgets/entete_application.dart';
 import 'package:jdr_maker/src/app/widgets/interfaces/app_interface/app_interface.dart';
 import 'package:jdr_maker/src/domain/data/couleurs.dart';
-import 'package:jdr_maker/src/domain/enums/form_bouton_type.dart';
 import 'package:jdr_maker/src/domain/models/evenement_model.dart';
 import 'package:provider/provider.dart';
 
@@ -97,10 +96,12 @@ class _EvenementEditViewState extends State<EvenementEditView> {
                       ),
                     ],
                   ),
-                  FormBouton(
-                    boutonType: FormBoutonType.valider,
-                    alignement: Alignment.bottomRight,
-                    action: modifier,
+                  Align(
+                    alignment: Alignment.bottomRight,
+                    child: BoutonIcone(
+                      icone: Icons.done_rounded,
+                      action: modifier,
+                    ),
                   ),
                 ],
               ),