+++ lib/src/domain/enums/form_bouton_type.dart
@@ -1,8 +1,0 @@
-/// Type de formulaire bouton
-enum FormBoutonType {
-  ajouter,
-  valider,
-  fermer,
-  modifier,
-  supprimer,
-}